pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autocoder"
version = "0.1.0"
description = "URL-driven Playwright BDD test automation with local Phi-4 14B (Ollama). Auth-first, low-token, resumable."
requires-python = ">=3.11"
readme = "readme/README.md"
dependencies = [
    "playwright>=1.48.0",
    "pytest>=8.3.0",
    "pytest-bdd>=7.3.0",
    "pytest-playwright>=0.5.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0.2",
    "click>=8.1.7",
    "rich>=13.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
autocoder = "autocoder.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["autocoder*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]
